feature/이메일인증디버깅 및 폰트 수정
